requirements.txt
langgraph
langchain
langchain-anthropic
langchain-community
chromadb
sentence-transformers
anthropic
python-dotenv
pydantic
streamlit
tiktoken
plotly
pandas